Squashed commit focusing on compilation speed
Enable test for asEigenMatrix and member_view

Split code from group.h --> cpp file and tests --> group_test.h

Decrease header dependencies and split functionparser to separate target

Shared libfaunus for several CMake targets; Updated doxygen
documentation

Fix MPI build
