+++ .idea/misc.xml
@@ -5,6 +5,9 @@
     <excludeRoots>
       <file path="$PROJECT_DIR$/3rdparty" />
       <file path="$PROJECT_DIR$/_deps" />
+      <file path="$PROJECT_DIR$/cmake-build-debug" />
+      <file path="$PROJECT_DIR$/include" />
+      <file path="$PROJECT_DIR$/include/faunus/sasa" />
     </excludeRoots>
   </component>
 </project>
