+++ src/random.h
@@ -23,7 +23,14 @@ namespace Faunus {
         Random();
         void seed();
         double operator()(); //!< Double in uniform range [0,1)
-        int range(int, int); //!< Integer in uniform range [min:max]
+
+        /**
+         * @brief Integer in uniform range [min:max]
+         * @param min minimum value
+         * @param max maximum value
+         * @return random value in [min:max] range
+         */
+        int range(int, int);
 
         template<class Titer>
             Titer sample(Titer begin, Titer end) {
